requirements.txt
firebase-admin
pandas
python-dotenv
mysql-connector-python